Refactor plugin base class
